pyproject.toml
[tool.poetry]
name = "selenium-movie-picker"
version = "0.1.0"
description = ""
authors = ["Cody Bontecou <bontecouc@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
selenium = "^3.141.0"
IMDbPY = "^2021.4.18"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
